switch to uint32_t for session category type, seemed to not be present after converting to workspace

+++ src/audio2/cocoa/DeviceManagerAudioSession.cpp
@@ -172,11 +172,11 @@ void DeviceManagerAudioSession::activateSession()
 	mSessionIsActive = true;
 }
 
-UInt32	DeviceManagerAudioSession::getSessionCategory()
+uint32_t DeviceManagerAudioSession::getSessionCategory()
 {
 	UInt32 result;
 	audioSessionProperty( kAudioSessionProperty_AudioCategory, result );
-	return result;
+	return static_cast<uint32_t>( result );
 }
 
 } } // namespace audio2::cocoa
